+++ pysrc/bytewax/connectors/kafka/__init__.py
@@ -1,0 +1,49 @@
+"""Connectors for [Kafka](https://kafka.apache.org).
+
+Importing this module requires the
+[`confluent-kafka`](https://github.com/confluentinc/confluent-kafka-python)
+package to be installed.
+
+The input source returns a stream of `KafkaMessage`.
+See the docstring for its use.
+
+You can use `KafkaSource` and `KafkaSink` directly:
+
+>>> from bytewax.connectors.kafka import KafkaSource, KafkaSink
+>>> from bytewax import operators as op
+>>> from bytewax.dataflow import Dataflow
+>>>
+>>> brokers = ["localhost:1909"]
+>>> flow = Dataflow("example")
+>>> kinp = op.input("kafka-in", flow, KafkaSource(["localhost:1909"], ["in-topic"]))
+>>> processed = op.map("map", kinp, lambda x: (x.key, x.value))
+>>> op.output("kafka-out", processed, KafkaSink(brokers, "out-topic"))
+
+Or the custom operators:
+
+>>> from bytewax.connectors.kafka import operators as kop
+>>> from bytewax import operators as op
+>>> from bytewax.dataflow import Dataflow
+>>>
+>>> brokers = ["localhost:1909"]
+>>> flow = Dataflow("example")
+>>> kinp = kop.input("kafka-in", flow, brokers=brokers, topics=["in-topic"])
+>>> errs = op.inspect("errors", kinp.errs).then(op.raises, "crash-on-err")
+>>> processed = op.map("map", kinp.oks, lambda x: (x.key, x.value))
+>>> kop.output("kafka-out", processed, brokers=brokers, topic="out-topic")
+
+"""
+from . import operators, registry, serde
+from .message import KafkaSinkMessage, KafkaSourceMessage
+from .sink import KafkaSink
+from .source import KafkaSource
+
+__all__ = [
+    "KafkaSource",
+    "KafkaSink",
+    "KafkaSinkMessage",
+    "KafkaSourceMessage",
+    "operators",
+    "registry",
+    "serde",
+]